prevent 32-bit overflow in the hashtable code
git-svn-id: https://svn.code.sf.net/p/msieve/code/trunk@408 8c02911e-0617-4013-80d0-c3bcbcae04d5

+++ common/hashtable.c
@@ -135,7 +135,7 @@ void *hashtable_find(hashtable_t *h, void *blob,
 
 	offset = hashtable[hashval];
 	while (offset != 0) {
-		entry = match_array + offset * (blob_words + 1);
+		entry = match_array + (size_t)offset * (blob_words + 1);
 		for (i = 0; i < hash_words; i++) {
 			if (entry[i] != key[i])
 				break;
@@ -156,7 +156,9 @@ void *hashtable_find(hashtable_t *h, void *blob,
 
 		/* not found; add it */
 
-		entry = match_array + h->match_array_size * (blob_words + 1);
+		entry = match_array + 
+			(size_t)h->match_array_size * (blob_words + 1);
+
 		for (i = 0; i < hash_words; i++) {
 			entry[i] = key[i];
 		}
